Minor Changes to Db Tables. Added Data Generation project

+++ SuplementsShop/Przesylka.sql
@@ -2,7 +2,7 @@
 (
 	[Przesylka_Id] INT IDENTITY PRIMARY KEY, 
     [PotwierdzoneDostarczenie] BIT NOT NULL, 
-    [Wartosc] INT NOT NULL, 
+    [Wartosc] FLOAT NOT NULL, 
     [AdresDostawy] INT NOT NULL, 
     CONSTRAINT [FK_Przesylka_Adres] FOREIGN KEY ([AdresDostawy]) REFERENCES [Adres]([Adres_Id]), 
 )
